feat: added LoginForm and RegisterForm
