+++ views/auth/LoginView.vue
@@ -1,7 +1,6 @@
 <script setup>
-import { ref } from 'vue'
 import AppLayout from '@/components/layout/AppLayout.vue'
-const showPassword = ref(false)
+import LoginForm from '@/components/auth/LoginForm.vue'
 
 import { useDisplay } from 'vuetify'
 const { mobile } = useDisplay()
@@ -25,39 +24,7 @@ const { mobile } = useDisplay()
 
             <v-card-text class="bg-surface-light pt-4">
               <v-divider class="my-5"></v-divider>
-              <v-form fast-fail @submit.prevent>
-                <v-text-field
-                  label="Email"
-                  type="email"
-                  variant="outlined"
-                  prepend-inner-icon="mdi-email"
-                ></v-text-field>
-
-                <v-text-field
-                  label="Password"
-                  variant="outlined"
-                  prepend-inner-icon="mdi-lock"
-                  :type="showPassword ? 'text' : 'password'"
-                  :append-inner-icon="showPassword ? 'mdi-eye-off' : 'mdi-eye'"
-                  @click:append-inner="showPassword = !showPassword"
-                ></v-text-field>
-
-                <v-btn
-                  color="orange-darken-2"
-                  class="mt-2"
-                  type="submit"
-                  block
-                  prepend-icon="mdi-login"
-                  >Login</v-btn
-                >
-                <v-divider class="my-5"></v-divider>
-                <h4 class="text-center">
-                  Don't have account?
-                  <RouterLink class="text-primary" to="/register"
-                    >Click here to Register.</RouterLink
-                  >
-                </h4>
-              </v-form>
+              <LoginForm></LoginForm>
             </v-card-text>
           </v-card>
         </v-col>
